+++ tests/PluginTest.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * Unit test for the plugin
+ */
+declare(strict_types=1);
+
+namespace Tests;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class PluginTest
+ * @package Tests
+ */
+class PluginTest extends TestCase
+{
+    public function test()
+    {
+        $this->assertTrue(true);
+    }
+}
